refactor: fixture 패키지명을 dao로 변경
