+++ src/test/java/roomescape/dao/FakeReservationDao.java
@@ -1,11 +1,10 @@
-package roomescape.fixture;
+package roomescape.dao;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
-import roomescape.dao.ReservationRepository;
 import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
