+++ src/test/java/roomescape/dao/FakeReservationTimeDao.java
@@ -1,10 +1,9 @@
-package roomescape.fixture;
+package roomescape.dao;
 
 import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
-import roomescape.dao.ReservationTimeRepository;
 import roomescape.model.ReservationTime;
 
 public class FakeReservationTimeDao implements ReservationTimeRepository {
