+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -5,13 +5,13 @@ import java.time.LocalTime;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.dao.FakeReservationDao;
+import roomescape.dao.FakeReservationTimeDao;
 import roomescape.dao.ReservationRepository;
 import roomescape.dao.ReservationTimeRepository;
 import roomescape.dto.ReservationReqDto;
 import roomescape.dto.ReservationResDto;
 import roomescape.dto.ReservationTimeResDto;
-import roomescape.fixture.FakeReservationDao;
-import roomescape.fixture.FakeReservationTimeDao;
 
 import static org.assertj.core.api.Assertions.assertThat;
